Fri, Mar 24, 2023, 5:10 AM +00:00

+++ week4/filter-more/helpers.c
@@ -3,18 +3,18 @@
 
 // Convert image to grayscale
 // For each pixel, turn it into grayscale
-// To turn to grayscale,
+// Grayscale is calculate by approximating the avera
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
-    int gray;
+    float gray;
     for (int i = 0; i < height; i++)
     {
         for (int j = 0; j < width; j++)
         {
             gray = ceil((image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue)/3);
-            image[i][j].rgbtRed = gray;
-            image[i][j].rgbtGreen = gray;
-            image[i][j].rgbtBlue = gray;
+            image[i][j].rgbtRed = (int)gray;
+            image[i][j].rgbtGreen = (int)gray;
+            image[i][j].rgbtBlue = (int)gray;
         }
     }
     return;
